Update patch set 1
Patch Set 1:

Hm.. I'm not sure Drive uses HTMLViewer necessarily intentionally, but HTMLViewer is the app that responds to the intent to view a plain text file. Feels like the correct way to fix this is to extend HTMLViewer with a way for apps to specify the correct encoding for the data they are sending, perhaps it should be set in the Bundle on the Intent and read back by HTMLViewer.

That, of course would require Drive to also change their code, but this bug seems like a motivator for that to happen.

Patch-set: 1