Modernize HTMLViewer, offer to decompress.
Spruce up the UI to match Material spec.  We can also load file://
content directly instead of forcing through a ContentProvider.

Since we're using this to show license content, try decompressing
files that end in ".gz", but fall back to loading normally if it's
not compressed.  Let callers provide a window title.

Bug: 18376908
Change-Id: I59ce02fb17fbe157239d365417a1a1e88e2a18d4