Update patch set 2
Patch Set 2:

Yes, I understand what you are trying to do. My point is that the code you have copied does a great deal of nontrivial work to determine whether to send the link out as an intent to another app or not, when you can simply return true from the function to make WebView send it out as an intent unconditionally, assuming this doesn't break anything else. The change you've written here will still, in some cases, follow the link inside HTMLViewer because it still returns false in some conditions.

Patch-set: 2