Reload data from an external storage after the permission is granted
HTMLViewer tries to load data from an external storage even if
the permission is denied. It just shows up 'Webpage not available',
so users can't notice that this is the permission problem.
It would be better to check if the file is in external storage
and show dialog to check permission. HTMLViewer gets closed and
shows toast message if permission is denied. On the other hand,
it reloads data if permission is granted by a user.

Change-Id: I26ad05136194cb36d4a17d3a8ab1a3fd2b5dfeb5
Signed-off-by: Euisang Lim <eui-sang.lim@samsung.com>