Update patch set 6
Patch Set 6:

As in, there can be more than one external storage directory? I didn't know that. My model was that /sdcard would be a symlink to an OEM-chosen directory.

It sounds like we shouldn't try to detect external storage, then. Just keep the scheme-based check. Sorry for all the back-and-forth, Euisang.

Patch-set: 6
Label: Verified=0