Update patch set 1
Patch Set 1:

OK, that sounds reasonable.

The documentation says this should be added as a child of the <application> - there's only one activity in HTMLViewer so maybe that's equivalent at the moment, but it seems like we'd want it to be in <application> anyway since that's what other apps do.

Patch-set: 1
Label: Code-Review=0